Move out placementRequirementsRow
In order to break dependency cycles
