+++ server/utils/placementRequests/matchingInformationSummaryList.test.ts
@@ -1,7 +1,8 @@
 import { placementRequestDetailFactory, premisesFactory } from '../../testutils/factories'
-import { mapSearchParamCharacteristicsForUi } from '../matchUtils'
-import { matchingInformationSummary, placementRequirementsRow } from './matchingInformationSummaryList'
+import { matchingInformationSummary } from './matchingInformationSummaryList'
 import { getPreferredApsFromApplication } from './getPreferredApsFromApplication'
+import { preferredApsRow } from './preferredApsRow'
+import { placementRequirementsRow } from './placementRequirementsRow'
 
 jest.mock('./getPreferredApsFromApplication')
 
@@ -65,36 +66,6 @@ describe('matchingInformationSummaryList', () => {
           placementRequirementsRow(placementRequest, 'essential'),
           placementRequirementsRow(placementRequest, 'desirable'),
         ],
-      })
-    })
-  })
-
-  describe('placementRequirementsRow', () => {
-    it('returns a list of desirable placement requirements in sentence case', () => {
-      const placementRequest = placementRequestDetailFactory.build()
-
-      expect(placementRequirementsRow(placementRequest, 'desirable')).toEqual({
-        key: {
-          text: `Desirable Criteria`,
-        },
-        value: {
-          html: mapSearchParamCharacteristicsForUi(placementRequest.desirableCriteria),
-        },
-      })
-    })
-  })
-
-  describe('placementRequirementsRow', () => {
-    it('returns a list of essential placement requirements in sentence case', () => {
-      const placementRequest = placementRequestDetailFactory.build()
-
-      expect(placementRequirementsRow(placementRequest, 'essential')).toEqual({
-        key: {
-          text: `Essential Criteria`,
-        },
-        value: {
-          html: mapSearchParamCharacteristicsForUi(placementRequest.essentialCriteria),
-        },
       })
     })
   })
